Address character chat phase 6 review comments
